+++ cdp-langchain/examples/chatbot/wallet_data.txt.mainnet.bak
@@ -1,0 +1,1 @@
+{"walletId":"3ffeb5c8-5867-4d3a-9dfc-6638c983745c","seed":"f38482c904becf85ecb3a2b8ed31f3643852bbe367f52ca846f24fbe3d37e8ad","defaultAddressId":"0xBE345D76092Ff450e056390aec917A122cE7f7a4"}
