+++ cdp-langchain/examples/chatbot/wallet_data.txt.testnet.bak
@@ -1,0 +1,1 @@
+{"walletId":"fd811d4c-b8bf-4284-a984-c4698589d868","seed":"ec7aded9ea29aed26fc8314e0110903375ada666aa5bf6e6d84bf88a37e0ae64","defaultAddressId":"0x252A265cae1149b7FDf92754D1D3B2bE21466878"}
